requirements.txt
Cython
Flask
Flask-Cors
Flask-PyMongo
Flask-RESTful
Jinja2
MarkupSafe
PyJWT
PyYAML
Werkzeug
click
gunicorn
itsdangerous
pip
pymssql-linux
setuptools
six